pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "github-pr-comment-agent"
version = "0.1.0"
description = "Intelligent MCP agent for managing GitHub PR comments with Claude"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Mark Johnson", email = "mark@enverus.com"}
]
keywords = ["github", "pr", "comments", "claude", "mcp", "agent", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Version Control :: Git",
]

dependencies = [
    "anthropic>=0.40.0",
    "mcp>=1.0.0",
    "PyGithub>=2.1.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "gitpython>=3.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
pr-comment-agent = "cli.main:cli"

[project.urls]
Homepage = "https://github.com/MarkEnverus/claude-mcp-agent-github-comments"
Repository = "https://github.com/MarkEnverus/claude-mcp-agent-github-comments"
Issues = "https://github.com/MarkEnverus/claude-mcp-agent-github-comments/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["mcp_server*", "agent*", "cli*"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"test_*.py" = ["E402"]  # Test scripts need to load .env before imports
"apply_fixes_pr72.py" = ["E402"]  # Scripts need to load .env before imports
"examples/*.py" = ["E402"]  # Examples need to load .env before imports

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "github.*",
    "rich.*",
    "click.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=mcp_server",
    "--cov=agent",
    "--cov=cli",
    "--cov-report=term-missing",
    "--cov-report=html",
]
asyncio_mode = "auto"
